+++ Sources/NX10CoreSDK/Services/Telemetry/Models/TelemetryModels.swift
@@ -21,7 +21,7 @@ public struct DeviceTypePayload: Codable {
 }
 
 public struct KeyboardMetricsSummary: Codable {
-    public init(typingSpeedWpm: Int, backspaceCount: Int, erasedTextLength: Int, averageHoldTimeMs: Int64, flightTimesMs: [Int64], totalKeyPresses: Int) {
+    public init(typingSpeedWpm: Int, backspaceCount: Int, erasedTextLength: Int, averageHoldTimeMs: Double, flightTimesMs: [Double], totalKeyPresses: Int) {
         self.typingSpeedWpm = typingSpeedWpm
         self.backspaceCount = backspaceCount
         self.erasedTextLength = erasedTextLength
@@ -33,20 +33,20 @@ public struct KeyboardMetricsSummary: Codable {
     public let typingSpeedWpm: Int
     public let backspaceCount: Int
     public let erasedTextLength: Int
-    public let averageHoldTimeMs: Int64
-    public let flightTimesMs: [Int64]
+    public let averageHoldTimeMs: Double
+    public let flightTimesMs: [Double]
     public let totalKeyPresses: Int
 }
 
 public struct MotionSample: Codable {
-    public init(timestampMs: Int64, x: Double, y: Double, z: Double) {
+    public init(timestampMs: Double, x: Double, y: Double, z: Double) {
         self.timestampMs = timestampMs
         self.x = x
         self.y = y
         self.z = z
     }
     
-    public let timestampMs: Int64
+    public let timestampMs: Double
     public let x: Double
     public let y: Double
     public let z: Double
@@ -96,7 +96,7 @@ public struct GeneralTouchSample: Codable {
     /// Velocity components in UIKit points / second (0 when unavailable).
     public let velocityX:   Double
     public let velocityY:   Double
-    public let timestampMs: Int64
+    public let timestampMs: Double
     
     public init(touchId: String,
                 touchType: TouchType,
@@ -106,7 +106,7 @@ public struct GeneralTouchSample: Codable {
                 radiusMm: Double,
                 velocityX: Double = 0,
                 velocityY: Double = 0,
-                timestampMs: Int64) {
+                timestampMs: Double) {
         self.touchId     = touchId
         self.touchType   = touchType
         self.touchObject = touchObject
@@ -137,8 +137,8 @@ public struct GeneralTouchSample: Codable {
 public struct KbStateSample: Codable {
     /// "down" when the keyboard appeared; "up" when it disappeared.
     public let state:       String
-    public let timestampMs: Int64
-    public init(state: String, timestampMs: Int64) {
+    public let timestampMs: Double
+    public init(state: String, timestampMs: Double) {
         self.state = state; self.timestampMs = timestampMs
     }
 }
@@ -146,8 +146,8 @@ public struct KbStateSample: Codable {
 /// Records characters erased by a single backspace — maps to "text-del" V2 event.
 public struct TextDelSample: Codable {
     public let erasedLength: Int
-    public let timestampMs:  Int64
-    public init(erasedLength: Int, timestampMs: Int64) {
+    public let timestampMs:  Double
+    public init(erasedLength: Int, timestampMs: Double) {
         self.erasedLength = erasedLength; self.timestampMs = timestampMs
     }
 }
@@ -155,8 +155,8 @@ public struct TextDelSample: Codable {
 /// Records a text correction event — maps to the "text-cor" V2 event.
 public struct TextCorSample: Codable {
     public let correction:  String
-    public let timestampMs: Int64
-    public init(correction: String, timestampMs: Int64) {
+    public let timestampMs: Double
+    public init(correction: String, timestampMs: Double) {
         self.correction = correction; self.timestampMs = timestampMs
     }
 }
@@ -165,8 +165,8 @@ public struct TextCorSample: Codable {
 public struct ScreenEventSample: Codable {
     /// "lock" or "unlock".
     public let event:       String
-    public let timestampMs: Int64
-    public init(event: String, timestampMs: Int64) {
+    public let timestampMs: Double
+    public init(event: String, timestampMs: Double) {
         self.event = event; self.timestampMs = timestampMs
     }
 }
